requirements.txt
pillow>=10.0.0
matplotlib>=3.7.1 